renamed bingoboard to mbti
